refactor(foundation): fix circular dependency
